requirements.txt
imagecodecs==2024.12.30 # optional, for zip compression in tifffile lzw compression
numpy==2.0.2
opencv-python==4.12.0.88
pillow==11.3.0
tifffile==2024.8.30
